% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBinom.R
\docType{data}
\name{BinomData}
\alias{BinomData}
\title{This is data to be included in my package}
\format{An object of class \code{numeric} of length 30.}
\usage{
BinomData
}
\description{
This is data to be included in my package
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q1p7b.txt}
}
\author{
Jake Necklen \email{neckl004@umn.edu}
}
\keyword{datasets}
