% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosecol.R
\name{choosecol}
\alias{choosecol}
\title{Filter for a specific column in a matrix}
\usage{
choosecol(data, colname)
}
\arguments{
\item{data}{user's dataset}

\item{colname}{name of column to filter for}
}
\value{
matrix with filtered column
}
\description{
Filter for a specific column in a matrix
}
\examples{
#' choosecol(x, "gender")
}
