% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFarm.R
\docType{data}
\name{FarmData}
\alias{FarmData}
\title{This is data to be included in my package}
\format{An object of class \code{data.frame} with 9597 rows and 1754 columns.}
\usage{
FarmData
}
\description{
This is data to be included in my package
}
\references{
\url{http://users.stat.umn.edu/~almquist/3811_examples/data.csv}
}
\author{
Jake Necklen \email{neckl004@umn.edu}
}
\keyword{datasets}
