% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGrowth.R
\docType{data}
\name{GrowthData}
\alias{GrowthData}
\title{This is data to be included in my package}
\format{An object of class \code{data.frame} with 93 rows and 33 columns.}
\usage{
GrowthData
}
\description{
This is data to be included in my package
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/growth.csv}
}
\author{
Jake Necklen \email{neckl004@umn.edu}
}
\keyword{datasets}
