% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Finds the MLE based on given data, distribution and interval}
\usage{
loglikelihood(dataset, distribution, interval)
}
\arguments{
\item{dataset}{dataset to use}

\item{distribution}{distribution function}

\item{interval}{interval to search}
}
\value{
MLE
}
\description{
Finds the MLE based on given data, distribution and interval
}
\examples{
b <- scan(url("http://www.stat.umn.edu/geyer/3701/data/q1p7b.txt"))
binom = function(theta, b) {dbinom(b, 20, prob = 1 / (1 + exp(- theta)), log = TRUE)}
loglikelihood(b, binom, c(-100,100))
}
