% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMatrix.R
\docType{data}
\name{MatrixData}
\alias{MatrixData}
\title{This is data to be included in my package}
\format{An object of class \code{character} of length 1.}
\usage{
MatrixData
}
\description{
This is data to be included in my package
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q2p3.rda}
}
\author{
Jake Necklen \email{neckl004@umn.edu}
}
\keyword{datasets}
