% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myapply.R
\name{myapply}
\alias{myapply}
\title{Simulates the R function apply}
\usage{
myapply(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{data vector}

\item{MARGIN}{number (1=row-wise, 2=col-wise)}

\item{FUN}{function to apply to data vector}
}
\value{
vector
}
\description{
Simulates the R function apply
}
\examples{
myapply(matrix(rnorm(20),nrow=4,ncol=5),2,mean)

}
