% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\title{Calculate weighted mean, variance and standard deviation}
\usage{
probability(d)
}
\arguments{
\item{d}{data matrix}
}
\value{
list
}
\description{
Calculate weighted mean, variance and standard deviation
}
\examples{
probs=runif(10)
probability(list(x = runif(10), p=(probs/sum(probs))))

}
