% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probGIEMO.R
\name{probGIEMO}
\alias{probGIEMO}
\title{Calculate weighted mean, var, sd with GIEMO}
\usage{
probGIEMO(d)
}
\arguments{
\item{d}{data vector}
}
\value{
list
}
\description{
Calculate weighted mean, var, sd with GIEMO
}
\examples{
probs=runif(10)
probGIEMO(list(x = runif(10), p=(probs/sum(probs))))

}
