% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProbabilities.R
\docType{data}
\name{ProbsData}
\alias{ProbsData}
\title{This is data to be included in my package}
\format{An object of class \code{data.frame} with 30 rows and 2 columns.}
\usage{
ProbsData
}
\description{
This is data to be included in my package
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q1p4.txt}
}
\author{
Jake Necklen \email{neckl004@umn.edu}
}
\keyword{datasets}
