% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Filter for a specific column in a matrix}
\usage{
select(data, colname)
}
\arguments{
\item{data}{user's dataset}

\item{colname}{name of column to filter for}
}
\value{
matrix with filtered column
}
\description{
Filter for a specific column in a matrix
}
\examples{
#' select(x, "gender")
}
