% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardCols.R
\name{standardCols}
\alias{standardCols}
\title{Standardizes the columns of a given matrix}
\usage{
standardCols(mat)
}
\arguments{
\item{mat}{data matrix}
}
\value{
matrix (standardized)
}
\description{
Standardizes the columns of a given matrix
}
\examples{
standardCols(matrix(rnorm(20), nrow=5, ncol=4))

}
