% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted}
\alias{weighted}
\title{Calculate weighted mean, variance and standard deviation}
\usage{
weighted(d)
}
\arguments{
\item{d}{data matrix}
}
\value{
list
}
\description{
Calculate weighted mean, variance and standard deviation
}
\examples{
probs=runif(10)
weighted(list(x = runif(10), p=(probs/sum(probs))))

}
