% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedGIEMO.R
\name{weightedGIEMO}
\alias{weightedGIEMO}
\title{Calculate weighted mean, var, sd with GIEMO}
\usage{
weightedGIEMO(d)
}
\arguments{
\item{d}{data vector}
}
\value{
list
}
\description{
Calculate weighted mean, var, sd with GIEMO
}
\examples{
probs=runif(10)
weightedGIEMO(list(x = runif(10), p=(probs/sum(probs))))

}
