% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwrapper.R
\name{ggwrapper}
\alias{ggwrapper}
\title{Make a plot with user's x and y variables from a dataset}
\usage{
ggwrapper(data, X, Y)
}
\arguments{
\item{data}{user's dataset}

\item{X}{independent variable}

\item{Y}{dependent variable}
}
\value{
plot
}
\description{
Make a plot with user's x and y variables from a dataset
}
\examples{
d = read.csv("http://www.stat.umn.edu/geyer/3701/data/growth.csv")
ggwrapper(d, d[,1], d[,2])
}
