% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Finds the MLE based on given data, distribution and interval}
\usage{
loglikelihood(dataset, distribution, interval)
}
\arguments{
\item{dataset}{dataset to use}

\item{distribution}{distribution function}

\item{interval}{interval to search}
}
\value{
MLE
}
\description{
Finds the MLE based on given data, distribution and interval
}
\examples{
#' loglikelihood(x, binomial, c(-100,100))
}
